+++ src/Analyzers/CSharp/Analysis/RequiredConfigOptionNotSetAnalyzer.cs
@@ -53,6 +53,7 @@ namespace Roslynator.CSharp.Analysis
                     Validate(ref context, compilationOptions, options, Flags.AddOrRemoveParenthesesFromConditionInConditionalOperator, ref flags, DiagnosticRules.AddOrRemoveParenthesesFromConditionInConditionalOperator, ConfigOptions.ConditionalOperatorConditionParenthesesStyle);
                     Validate(ref context, compilationOptions, options, Flags.ConfigureAwait, ref flags, DiagnosticRules.ConfigureAwait, ConfigOptions.ConfigureAwait);
                     Validate(ref context, compilationOptions, options, Flags.IncludeParenthesesWhenCreatingNewObject, ref flags, DiagnosticRules.IncludeParenthesesWhenCreatingNewObject, ConfigOptions.ObjectCreationParenthesesStyle);
+                    Validate(ref context, compilationOptions, options, Flags.NormalizeFormatOfEnumFlagValue, ref flags, DiagnosticRules.NormalizeFormatOfEnumFlagValue, ConfigOptions.EnumFlagValueStyle);
                     Validate(ref context, compilationOptions, options, Flags.NormalizeNullCheck, ref flags, DiagnosticRules.NormalizeNullCheck, ConfigOptions.NullCheckStyle);
                     Validate(ref context, compilationOptions, options, Flags.UseAnonymousFunctionOrMethodGroup, ref flags, DiagnosticRules.UseAnonymousFunctionOrMethodGroup, ConfigOptions.UseAnonymousFunctionOrMethodGroup);
                     Validate(ref context, compilationOptions, options, Flags.UseBlockBodyOrExpressionBody, ref flags, DiagnosticRules.UseBlockBodyOrExpressionBody, _useBlockBodyOrExpressionBodyOptions);
@@ -109,6 +110,7 @@ namespace Roslynator.CSharp.Analysis
             ConfigureAwait,
             DocCommentSummaryStyle,
             IncludeParenthesesWhenCreatingNewObject,
+            NormalizeFormatOfEnumFlagValue,
             NormalizeNullCheck,
             NormalizeUsageOfInfiniteLoop,
             UseAnonymousFunctionOrMethodGroup,