+++ src/Common/ConfigOptionKeys.Generated.cs
@@ -21,6 +21,7 @@ namespace Roslynator
         public const string ConfigureAwait                                    = "roslynator_configure_await";
         public const string DocCommentSummaryStyle                            = "roslynator_doc_comment_summary_style";
         public const string EmptyStringStyle                                  = "roslynator_empty_string_style";
+        public const string EnumFlagValueStyle                                = "roslynator_enum_flag_value_style";
         public const string EnumHasFlagStyle                                  = "roslynator_enum_has_flag_style";
         public const string EqualsTokenNewLine                                = "roslynator_equals_token_new_line";
         public const string InfiniteLoopStyle                                 = "roslynator_infinite_loop_style";